Start to encapsulate the whole handler loading workflow
